EK-108: by cszablew, fixed <img> inline css, renamed files to PascalCaseX
